subsys/coredump: Fix big-endian support
Previously, compiling the coredump_core.c file generated errors like
these on SPARC (big-endian with 32-bit pointers):

  subsys/debug/coredump/coredump_core.c: In function
  'coredump_memory_dump':

  include/zephyr/sys/byteorder.h:34:42: error: right shift count >=
  width of type [-Werror=shift-count-overflow]
     34 | #define __bswap_64(x) ((uint64_t) ((((x) >> 56) & 0xff) | \

  include/zephyr/sys/byteorder.h:266:30: note: in expansion of macro
  '__bswap_64'
    266 | #define sys_cpu_to_le64(val) __bswap_64(val)

  subsys/debug/coredump/coredump_core.c:180:27: note: in expansion of
  macro 'sys_cpu_to_le64'
    180 |                 m.start = sys_cpu_to_le64(start_addr);

This commit provides the __bswap_64 macro parameters as uint64_t, which
eliminates the error.

+++ subsys/debug/coredump/coredump_core.c
@@ -177,8 +177,8 @@ void coredump_memory_dump(uintptr_t start_addr, uintptr_t end_addr)
 	m.hdr_version = COREDUMP_MEM_HDR_VER;
 
 	if (sizeof(uintptr_t) == 8) {
-		m.start	= sys_cpu_to_le64(start_addr);
-		m.end = sys_cpu_to_le64(end_addr);
+		m.start	= sys_cpu_to_le64((uint64_t) start_addr);
+		m.end = sys_cpu_to_le64((uint64_t) end_addr);
 	} else if (sizeof(uintptr_t) == 4) {
 		m.start	= sys_cpu_to_le32(start_addr);
 		m.end = sys_cpu_to_le32(end_addr);
